refactor, added to readme
