--- utils/README.md
@@ -1,54 +1,0 @@
-
-# Awesome Project
-
-<img src="https://img.shields.io/badge/License-MIT-green" alt="License" />
-
-## A description of an awesome project
----
-## Table of Contents
-- [Installation](#installation)
-- [Usage](#usage)
-- [Contribute](#contribute)
-- [Tests](#tests)
-- [Questions](#questions)
-- [License](#license)
-
----
-
-## Installation
-​
-Some installation instructions
-
----
-
-## Usage
-
-some usage instructions
-
-
----
-
-## Contribute
-
-contribute!
-
----
-
-## License
-
-This projet is licensed under the terms of MIT license.
-
----
-
-## Tests
-
-here are them tests yo
-
----
-
-## Questions
-
-If you have any questions about this repo,
-open an issue or contact jdmartinez@something.com
-
-
